Move convertNodeInfoToLegacyFormat() to P2P misc utils
